soldier_assistant/requirements.txt
paho-mqtt
pvporcupine
vosk
pyaudio
python-dotenv
sounddevice
noisereduce
python-dotenv
adapt-parser